database and bip moved commons init
