+++ app/channel/base/__init__.py
@@ -3,7 +3,7 @@ from app.commons.constants.globals import *
 from app.commons.log import get_logger
 from config import APP
 from app.commons.service import get_report_file_info
-from .core import *
+from .__core__ import *
 
 logger = get_logger(__name__)
