+++ jobs.py
@@ -5,13 +5,13 @@ import time
 
 from config import APP
 from bipwrapper.type.poll_type import PollType
+from app.commons import bip, database
 from app.commons.utils import *
 from app.commons.log import get_logger
 from app.commons.constants.globals import Globals
 from app.commons.constants.message import Message
 from app.commons.constants.command import Command
 from app.commons.cache import Cache, Keys
-from app.commons.service import bip, database
 
 logger = get_logger(__name__)
